Fix linter

+++ src/cellrank/__init__.py
@@ -1,7 +1,7 @@
 from importlib import metadata
 
-import scipy.sparse as sp
 import numpy as np
+import scipy.sparse as sp
 
 from cellrank import datasets, estimators, kernels, logging, models, pl
 from cellrank._utils._lineage import Lineage
@@ -20,6 +20,6 @@ except ImportError:
 # pygam uses `.A`
 sp.spmatrix.A = property(lambda self: self.toarray())
 # pygam/utils.py:649: in b_spline_basis
-np.int = int
+int = int
 
 del metadata, md, sp, np
